% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertor.R
\name{convertor}
\alias{convertor}
\title{Convert a cleaned dataset into a model.matrix}
\usage{
convertor(data, response, x.indices)
}
\arguments{
\item{data}{A well-cleaned raw \code{data.frame} that you would like to convert into a model.matrix as an input for the glmnet package.}

\item{response}{The column location of y in the dataset.}

\item{x.indices}{The columns of Xs that you would like to convert into a \code{model.matrix}.}
}
\value{
a list with elements:
         \item{response}{the original response}
         \item{X.model.matrix}{the \code{model.matrix} for the predictors with the intercept column removed}
}
\description{
This function converts a raw dataset into a user defined model.matrix.
}
\examples{

data <- data.frame(matrix(rnorm(100*5),100,5))
#Assume y is on the first coordinate
#and the rest are Xs:
model_mat <- convertor(data = data, response = 1, x.indices = seq(2,5))


}
\author{
Mokyo Zhou
}
