% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitter.R
\name{datasplit}
\alias{datasplit}
\title{Split the data into a training set and a test set}
\usage{
datasplit(data = data, splits = splits, random = FALSE)
}
\arguments{
\item{data}{A well-cleaned \code{data.frame} that you would like to split into a training set and a test set.}

\item{splits}{A two-element vector with the first element specifying the training proportion and the second being the test proportion.}

\item{random}{If TRUE, uses a random seed so you will get a different training and testing set everytime you run this. If FALSE, the draws are based on the seed (12345678). Default is random = FALSE.}
}
\value{
a list with elements:
         \item{train}{the training set}
         \item{test}{the test set}
}
\description{
This function splits the \code{data.frame} into a training set and a test set with user defined proportions.
}
\examples{

data <- data.frame(matrix(rnorm(100*5),100,5))
traintest <- datasplit(data = data, splits = c(0.8,0.2), random = FALSE)
training <- traintest[[1]]
testing <- traintest[[2]]



}
\author{
Mokyo Zhou
}
