% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en.robust.R
\name{en.robust}
\alias{en.robust}
\title{Internal lassoenet function}
\usage{
en.robust(x = x, y = y, alpha.seq = alpha.seq, err.curves = err.curves,
  result.matrix = result.matrix, parallel = parallel)
}
\arguments{
\item{x}{A model.matrix for the predictors.\cr}

\item{y}{A vector of response values.\cr}

\item{alpha.seq}{The alpha sequence for the Elastic Net}

\item{err.curves}{The number of error curves to be fitted. Default is 0.\cr}

\item{result.matrix}{A matrix for storing results.\cr}

\item{parallel}{Parallelisation}
}
\value{
A vector of results for the best Elastic Net model along with some plots of the error curves, under the condition where the full dataset has been used for modelling. The return from this function will enter \code{\link{prediction_ElasticNet}}.
}
\description{
Internal lassoenet function
}
\section{Details}{
 These are not intended for use by users. This function is one of the main engines for the Elastic Net computation. This function is used when the user does not want to split the dataset into
 a training and a test set. This together with the function \code{\link{no_split_en}} form the computation operator for the Elastic Net when using the full dataset. The return from this function will enter \code{\link{prediction_ElasticNet}} for futher wrapping.
}

\author{
Mokyo Zhou
}
