% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HitsNumber.R
\name{HitsNumber}
\alias{HitsNumber}
\title{Calculate how many times the lasso selected each variable}
\usage{
HitsNumber(regr_result, variable_list)
}
\arguments{
\item{regr_result}{A dataframe containing coefficients for each exposure for each run}

\item{variable_list}{A data.frame containing a list of exposures for which regressions will be run, first column must contain exposure names}
}
\value{
A dataframe with number of hits for each exposure variable
}
\description{
Calculate how many times the lasso selected each variable
}
