% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lasso.R
\name{Lasso}
\alias{Lasso}
\title{Lasso with negative binomial outcome repeated for each imputed dataset}
\usage{
Lasso(
  input_data,
  sdq,
  variable_list,
  conf_list,
  penalty_factor,
  path,
  file_name
)
}
\arguments{
\item{input_data}{A dataframe containing a complete stacked dataset}

\item{sdq}{A string defining the output variable ("h_sdq_external" or "h_sdq_internal")}

\item{variable_list}{A character vector containing a list of exposures to be included in the lasso}

\item{conf_list}{A character vector listing confounders to be used in the model}

\item{penalty_factor}{An integer vector with 1 meaning variables to be shrunk and 0 variables that are forced not to be shrunk (the number is a sum of variables and contrasts for categorical variables)}

\item{path}{A path for saving the file}

\item{file_name}{File name (without extension)}
}
\value{
A dataframe containing coefficients for each variable for each run
}
\description{
Lasso with negative binomial outcome repeated for each imputed dataset
}
