% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internallassoenetsim1.R
\name{MC.result}
\alias{MC.result}
\title{Internal lassoenet simulation function}
\usage{
MC.result(vector.result)
}
\arguments{
\item{vector.result}{A vector of simluation results}
}
\description{
Internal lassoenet simulation function
}
\section{Details}{
 This function calculates the averaged simulation prediction errors and it calls the function
\code{\link{SE.resample}} for standard error computation.
}

\author{
Mokyo Zhou
}
