% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_split_en.R
\name{no_split_en}
\alias{no_split_en}
\title{Internal lassoenet function}
\usage{
no_split_en(err.curves = 0, x = x, y = y, step.size = 0,
  type.lambda = type.lambda, parallel = parallel)
}
\arguments{
\item{err.curves}{The number of error curves to be fitted. Default is 0.\cr}

\item{x}{A model.matrix for the predictors.\cr}

\item{y}{A vector of response values.\cr}

\item{step.size}{Step size of the alpha grid.\cr}

\item{type.lambda}{Either "lambda.min" or "lambda.1se", default is "lambda.min.\cr}

\item{parallel}{Parallelisation}
}
\value{
A vector of results for the best Elastic Net model under the condition where the full dataset has been used for modelling. The return from this function will enter \code{\link{prediction_ElasticNet}}.
}
\description{
Internal lassoenet function
}
\section{Details}{
 These are not intended for use by users. This function is one of the main engines for the Elastic Net computation. This function is used when the user does not want to split the dataset into
 a training and a test set. This together with the function \code{\link{en.robust}} form the computation operator for the Elastic Net when using the full dataset. The return from this function will enter \code{\link{prediction_ElasticNet}} for futher wrapping.
}

\author{
Mokyo Zhou
}
