% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_split_lasso.R
\name{no_split_lasso}
\alias{no_split_lasso}
\title{Internal lassoenet function}
\usage{
no_split_lasso(err.curves = 0, x = x, y = y, type.lambda = type.lambda,
  al = 1, parallel = parallel)
}
\arguments{
\item{err.curves}{The number of error curves to be fitted. Default is 0.\cr}

\item{x}{A model.matrix for the predictors.\cr}

\item{y}{A vector of response values.\cr}

\item{type.lambda}{Either "lambda.min" or "lambda.1se", default is "lambda.min.\cr}

\item{al}{1 for the Lasso and 0 for ridge regression.\cr}

\item{parallel}{Parallelisation}
}
\value{
A vector of results for the best Lasso model under the condition where the full dataset has been used for modelling. The return from this function will enter \code{\link{prediction_Lasso}}.
}
\description{
Internal lassoenet function
}
\section{Details}{
 These are not intended for use by users. This function is one of the main engines for the Lasso computation. This function is used when the user does not want to split the dataset into
 a training and a test set. This together with the function \code{\link{split_lasso}} form the overall computation operator for the Lasso models. The return from this function will enter \code{\link{prediction_Lasso}} for futher wrapping.
}

\author{
Mokyo Zhou
}
