% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalised_pred.R
\name{penalised_pred}
\alias{penalised_pred}
\title{Internal lassoenet functions}
\usage{
penalised_pred(data = data, parallel = parallel, response = response,
  x.indices = x.indices, err.curves = err.curves,
  type.lambda = type.lambda)
}
\arguments{
\item{data}{A well-cleaned \code{data.frame}.\cr}

\item{parallel}{parallelisation.\cr}

\item{response}{The location of the response within the \code{data.frame}.\cr}

\item{x.indices}{The locations of the predictors withint the \code{data.frame}.\cr}

\item{err.curves}{The number of error curves to fit.\cr}

\item{type.lambda}{Either "lambda.min" or "lambda.1se".}
}
\value{
A vector of outputs and some plots related to the best model. The return from this function will enter the \code{\link{interactiver}} function.
}
\description{
Internal lassoenet functions
}
\section{Details}{
 These are not intended for use by users. This function is the overall wrapper for the prediction focused path. It takes returns from both the \code{\link{prediction_Lasso}} and \code{\link{prediction_ElasticNet}}
and put these through the comparison functions \code{\link{prediction.nonsplit.result}} and \code{\link{prediction.split.result}}. The return from this function will enter the \code{\link{interactiver}} function.
}

\author{
Mokyo Zhou
}
