% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotNoHits.R
\name{PlotNoHits}
\alias{PlotNoHits}
\title{Plot the number of the lasso hits per variable}
\usage{
PlotNoHits(no_hits, title, path, file_name)
}
\arguments{
\item{no_hits}{A dataframe with number of hits for each exposure variable}

\item{title}{A string defining the title of the plot}

\item{path}{A string defining the path for saving the file}

\item{file_name}{A string defining file name (without extension)}
}
\value{
A barplot of number of hits per variable
}
\description{
Plot the number of the lasso hits per variable
}
