% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.result2.R
\name{prediction.split.result}
\alias{prediction.split.result}
\title{Internal lassoenet functions}
\usage{
prediction.split.result(best.lasso.result = best.lasso.result,
  best.EN.result = best.EN.result, data = data, x.indices = x.indices,
  response = response, alph = alph, splits = splits,
  parallel = parallel)
}
\arguments{
\item{best.lasso.result}{A vector contains the outputs from the function \code{\link{prediction_Lasso}}.\cr}

\item{best.EN.result}{A vector contains the outputs from the function \code{\link{prediction_ElasticNet}}.\cr}

\item{data}{A well-cleaned \code{data.frame}.\cr}

\item{x.indices}{Locations of the predictors within the \code{data.frame}.\cr}

\item{response}{Location of the response within the \code{data.frame}.\cr}

\item{alph}{An internal argument.\cr}

\item{splits}{A two element vector with the first element being the training proportion and the second element being the testing proportion.\cr}

\item{parallel}{Parallelsation.\cr}
}
\value{
A vector of outputs and some plots related to the best model.
}
\description{
Internal lassoenet functions
}
\section{Details}{
 These are not intended for use by users.
This function provides an automatic comparison between the Lasso and the Elastic Net model based on the user's preferred prediction metric. The return from this function will then enter the \code{\link{penalised_pred}} function.\cr
}

\author{
Mokyo Zhou
}
