% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question.ad.R
\name{question.ad}
\alias{question.ad}
\title{Internal lassoenet functions}
\usage{
question.ad(data = data, response = response, x.indices = x.indices)
}
\arguments{
\item{data}{A well-cleaned \code{data.frame}.}

\item{response}{The location of the response within the \code{data.frame}.\cr}

\item{x.indices}{The locations of the predictors within the \code{data.frame}.\cr}
}
\description{
Internal lassoenet functions
}
\section{Details}{
 These are not intended for use by users.
This function is a part of the \code{\link{interactiver}} function.
This function asks a series of questions and provides pratical suggestions. It returns a collection of inputs for the Adaptive Lasso model.
}

\author{
Mokyo Zhou
}
