% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questioning.R
\name{questioning}
\alias{questioning}
\title{Internal lassoenet functions}
\usage{
questioning(data = data)
}
\arguments{
\item{data}{A well-cleaned \code{data.frame}.}
}
\description{
Internal lassoenet functions
}
\section{Details}{
 These are not intended for use by users.
This function is a part of the \code{\link{interactiver}} function.
This function asks a series of questions and provides pratical suggestions. It returns a collection of inputs for the Lasso and the Elastic Net model.
}

\author{
Mokyo Zhou
}
