% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.approach.R
\name{ridge.approach}
\alias{ridge.approach}
\title{Internal lassoenet function}
\usage{
ridge.approach(x = x, y = y, err.curves = err.curves,
  type.lambda = type.lambda, gamma.seq = c(0.5, 1, 2), CI = CI,
  dataa = dataa, parallel = parallel, xx.indices = xx.indices,
  B.rep = B.rep, significance = significance)
}
\arguments{
\item{x}{A \code{model.matrix} of the predictors.\cr}

\item{y}{A vector of response values for the model fitting.\cr}

\item{err.curves}{The number of error curves to be fitted. Default is 0.\cr}

\item{type.lambda}{Either "lambda.min" or "lambda.1se".}

\item{gamma.seq}{The gamma sequence to try.\cr}

\item{CI}{TRUE for residual bootstrapping. In this version is always TRUE.}

\item{dataa}{Your full \code{data.frame}}

\item{parallel}{Parallelisation}

\item{xx.indices}{Locations of the predictors within dataa.}

\item{B.rep}{The number of residual bootstrappings to do.}

\item{significance}{The significance level of the confidence intervals e.g. 100(1-\eqn{\alpha})\%.\cr}
}
\value{
A vector of outputs of the best Adpative Lasso model and the 100(1-\eqn{\alpha})\% confidence intervals for the point estimates from using residual bootstrapping.
}
\description{
Internal lassoenet function
}
\section{Details}{
 These are not intended for use by users. This function is one of the main engines for the Adaptive Lasso computation. This function is used when the weighting method "ridge" is selected.
This together with the function \code{\link{ridge.robust}} forms the computation operator for the Adaptive Lasso when using the weighting method "ridge". The return from this function will enter \code{\link{Adlasso}} for futher wrapping.
}

\author{
Mokyo Zhou
}
