% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internallassoenetsim2.R
\name{SE.resample}
\alias{SE.resample}
\title{Internal lassoenet simulation function}
\usage{
SE.resample(vector.result)
}
\arguments{
\item{vector.result}{A vector of simluation results}
}
\description{
Internal lassoenet simulation function
}
\section{Details}{
 This function calculates the standard error.
}

\author{
Mokyo Zhou
}
