% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adlassosimulation.R
\name{simulation.adlasso}
\alias{simulation.adlasso}
\title{Simulation studies for Lasso and Adpative Lasso}
\usage{
simulation.adlasso(n.resample = n.resample, n = n, coeff = coeff,
  matrix.option = 1, collinear = collinear, sig = sig, option = 2,
  parallel = FALSE)
}
\arguments{
\item{n.resample}{The number of simulation datasets to generate\cr}

\item{n}{The number of rows in the each \code{data.frame}}

\item{coeff}{A vector a true coefficients fixed acrossed the simulated datasets\cr}

\item{matrix.option}{1: Using an Exchangeable correlation matrix to simulate the predictors\cr
2: Using an Autoregressive correlation matrix to simulate the predictors\cr}

\item{collinear}{The correlation levels within the \code{matrix.option}}

\item{sig}{The model inherent error, the \eqn{\sigma^2}}

\item{option}{1: split the dataset according to \code{c(split.prop, 1 - split.prop)}\cr
2: Use the whole dataset. Note When \code{option = 2}, the \code{split.prop} will be ignored\cr
Note: When using \code{simulation.adlasso} please make sure "option = 2" under the current package.}

\item{parallel}{Parallelisation}
}
\value{
A list of elements:
\item{final.table}{A summary table contains: 1.The averaged number of incorrectly classified coefficients. \cr
                                             2.The proportion of times the each method correctly chooses the true model\cr}
\item{lasso}{A matrix full with simulation results for the Lasso method. Users can freely use this result matrix to obtain further insight\cr}
\item{Adpative Lasso}{A matrix full with simualtion results for the Adatpive Lasso method. Users can freely use this result matrix to obtain further insight\cr}
}
\description{
This function calls the data generating function \code{simulation.generation.data} and can be used
to conduct simulation studies for comparison between the Lasso and the Adaptive Lasso.
}
\section{Details}{
 The function is one of the core function for the simulation studies. The function supports comparison between the Lasso and the Adaptive Lasso in terms of making inferenes.
This function calls the function \code{simulation.generation.data} and thus, users can study different datasets of their liking. The function uses the coefficients from the Lasso fit to construct the
initial weights. The function investigates and compare between three different \eqn{\gamma}s e.g. c(0.5,1,2). The function provide a summary table for the simulation results. Furthermore, the function also
poduces visulisation of bias, variance and mean-squared error for the sampling distributions of the coefficients. Finally, the matrix containing the simulation iterations for each of these methods is also provided. Therefore, users are free to conduct further investigation.
}

\examples{
#number of simulated dataset 20, 400 rows, vector of true coefficient is c(0.4,0.4,2.5,2.5,0,0,0,
#0,0,0),using exchangeable correlation matrix, correlation level is 0.2 in the autoregressive
#matrix,model error is 2, using the whole dataset. No parallelisation.
simulation1 <- simulation.adlasso(n.resample = 20, n = 400, coeff = c(0.4,0.4,2.5,2.5,rep(0,6)),
matrix.option = 1, collinear = 0.2, sig = 2, parallel = FALSE,option=2)
}
\author{
Mokyo Zhou
}
