% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictionsimulation.R
\name{simulation.collinear}
\alias{simulation.collinear}
\title{Simulation studies for Lasso, Elastic Net and ridge}
\usage{
simulation.collinear(n.resample = n.resample, n = n, coeff = coeff,
  matrix.option = 2, collinear = collinear, sig = sig,
  split.prop = split.prop, step.alpha = step.alpha, option = 1,
  parallel = FALSE)
}
\arguments{
\item{n.resample}{The number of simulation datasets to generate\cr}

\item{n}{The number of rows in the each \code{data.frame}}

\item{coeff}{A vector a true coefficients fixed acrossed the simulated datasets\cr}

\item{matrix.option}{1: Using an Exchangeable correlation matrix to simulate the predictors\cr
2: Using an Autoregressive correlation matrix to simulate the predictors\cr}

\item{collinear}{The correlation levels within the \code{matrix.option}}

\item{sig}{The model inherent error, the \eqn{\sigma^2}}

\item{split.prop}{An element specifying the training proportion. Note the testing proportion will be 1 - the training proportion.}

\item{step.alpha}{The step size of the alpha grid for the Elastic Net.}

\item{option}{1: split the dataset according to \code{c(split.prop, 1 - split.prop)}\cr
2: Use the whole dataset. Note When \code{option = 2}, the \code{split.prop} will be ignored\cr
Note: When using \code{simulation.collinear} please make sure "option = 1" under the current package.}

\item{parallel}{Parallelisation}
}
\value{
A list of elements:
\item{final.table}{A summary table contains:
                                             1.The averaged external simulation prediction error for Lasso, Elastic Net and the Ridge\cr
                                             2.The standard error of the prediction error by using bootstrapping\cr
                                             3.The proportion of times the each method correctly chooses the true model\cr
                                             4.The averaged \eqn{\alpha}: 1 for Lasso, averaged simluated \eqn{\alpha} for the Elastic Net and 0 for the Ridge\cr}
\item{lasso.result}{A matrix full with simulation results for the Lasso method. Users can freely use this result matrix to obtain further insight\cr}
\item{EN.result}{A matrix full with simualtion results for the Elastic Net method. Users can freely use this result matrix to obtain further insight\cr}
\item{ridge.result}{A matrix full with simulation results for the ridge regression. Users can freely use this result matrix to botain further isnight\cr}
}
\description{
This function calls the data generating function \code{simulation.generation.data} and can be used
to conduct simulation studies for comparison between the Lasso, the Elastic Net and the ridge regression.
}
\section{Details}{
 The function is one of the core function for the simulation studies. The function supports comparison between the Lasso, the Elastic Net and the ridge regression.
This function calls the function \code{simulation.generation.data} and thus, users can study different datasets of their liking. The function provide a summary table for the simulation results.
The matrices containing the simulation iterations for the three methods are also provided. Therefore, users are free to conduct further investigation.
}

\examples{
#number of simulated dataset 20, 200 rows, vector of true coefficient is c(10,8,0,0,12,0,0,0,0,0),
#using autoregressive correlation matrix, correlation level is 0.2 in the autoregressive matrix,
#model error is 3, training proportion is 0.6, step size of the alpha grid is 0.2, splitting
#the dataset into a training and a testing set. No parallelisation.
simulation1 <- simulation.collinear(n.resample = 20, n = 200, coeff = c(10,8,0,0,12,0,0,0,0,0),
matrix.option = 2, collinear = 0.2, sig = 3,split.prop = 0.6, parallel = FALSE, step.alpha = 0.2,
option= 1)
}
\author{
Mokyo Zhou
}
