% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_en.R
\name{split_en}
\alias{split_en}
\title{Internal lassoenet function}
\usage{
split_en(err.curves = 0, x.train = x.train, y.train = y.train,
  x.test = x.test, y.test = y.test, step.size = 0,
  type.lambda = type.lambda, parallel = parallel)
}
\arguments{
\item{err.curves}{The number of error curves to be fitted. Default is 0.\cr}

\item{x.train}{A \code{model.matrix} of the preditors for the training data.\cr}

\item{y.train}{A vector of response values on the training set.\cr}

\item{x.test}{A \code{model.matrix} of the predictors for the testing data.\cr}

\item{y.test}{A vector of response values on the testing set.\cr}

\item{step.size}{The step size of the alpha sequence.\cr}

\item{type.lambda}{Either "lambda.min" or "lambda.1se", default is "lambda.min.\cr}

\item{parallel}{Parallelisation}
}
\value{
A vector of results for the best Elastic Net model under the condition where the full dataset has been splitted into a training and a testing set. The return from this function will enter \code{\link{prediction_ElasticNet}}.
}
\description{
Internal lassoenet function
}
\section{Details}{
 These are not intended for use by users. This function is one of the main engines for the Elastic Net computation. This function is used when the user would like to split the dataset into
 a training and a test set. This together with the function \code{\link{en.robust.split}} forms the computation operator for the Elastic Net when using the a traning and a testing set. The return from this function will enter \code{\link{prediction_ElasticNet}} for futher wrapping.
}

\author{
Mokyo Zhou
}
