% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_lasso.R
\name{split_lasso}
\alias{split_lasso}
\title{Internal lassoenet function}
\usage{
split_lasso(x.train = x.train, y.train = y.train, x.test = x.test,
  y.test = y.test, err.curves = 0, type.lambda = type.lambda,
  parallel = parallel)
}
\arguments{
\item{x.train}{A \code{model.matrix} of the preditors for the training data.\cr}

\item{y.train}{A vector of response values on the training set.\cr}

\item{x.test}{A \code{model.matrix} of the predictors for the testing data.\cr}

\item{y.test}{A vector of response values on the testing set.\cr}

\item{err.curves}{The number of error curves to be fitted. Default is 0.\cr}

\item{type.lambda}{Either "lambda.min" or "lambda.1se", default is "lambda.min.\cr}

\item{parallel}{Parallelisation}
}
\value{
A vector of results for the best Lasso model under the condition where the training set has been used to train and tune, and the testing set is used for obtaining the out of sample prediction error. The return from this function will enter \code{\link{prediction_Lasso}}.
}
\description{
Internal lassoenet function
}
\section{Details}{
 These are not intended for use by users. This function is one of the main engines for the Lasso computation. This function is used when the user would like to split the dataset into
 a training and a test set. This together with the function \code{\link{no_split_lasso}} form the overall computation operator for the Lasso models. The return from this function will enter \code{\link{prediction_Lasso}} for futher wrapping.
}

\author{
Mokyo Zhou
}
